refactor(cli): pass the state
