+++ crates/cli/src/lib.rs
@@ -5,6 +5,7 @@ use clap::Parser;
 use cli_args::CliArgs;
 use miette::set_panic_hook;
 use pacquet_diagnostics::enable_tracing_by_env;
+use state::State;
 
 pub async fn main() -> miette::Result<()> {
     enable_tracing_by_env();
